Allow webpage saving via persistent toolbar button in Chrome
Chrome now shows a persistent toolbar button for all extensions, even if
they only have page actions -- when there's no page action, it just
displays a grayed out version of the extension icon and shows the
standard toolbar button context menu on click.

This changes the connector to behave like Zotero for Firefox now does,
using the persistent button to allow webpage saving on all pages,
switching to translator-based saving when available. The menu option in
the content pane context menu now does the same thing as the toolbar
button.

It's not yet possible to choose other translators via a drop-down as in
Firefox. Chrome doesn't seem to allow you to customize the right-click
menu for a toolbar button, only to define a custom popup for left
clicks, so we'd have to either sacrifice one-click saving and show a
popup in all cases or show the other translator options in the content
pane context menu. I think we'd do the latter, unless we always wanted
to show a popup in order to allow choosing the target collection (but
that we could do the same way as the Firefox bookmark button, with a
second click on the same page showing a popup).

One current downside: webpage saving only works to Standalone, not the
server, and now there's a button for webpage saving showing nearly all
the time. (A clear error message appears in the popup, though, so it
might have the effect of encouraging people who weren't using Standalone
before to download it.) That said, it seems like we could easily make
this at least create webpage items on the server, even if it won't save
snapshots.