Support PDF saving via save button
Show "Save to Zotero (PDF)" in tooltip and send a flag to Standalone to
tell it to save the URL instead of saving the fake HTML wrapper page.
For now this requires a redownload on the Standalone side -- there might
be a better way to get the PDF data over without redownloading. (It uses
passed cookies, though, so gated PDFs should still work.)

This requires a new version of Standalone. Older versions will get the
HTML wrapper.

This is only hooked up for Chrome right now. This changes the
onTranslators signature to include a contentType, so Safari will need to
be updated accordingly.