Adds a secondary translator selector for Safari
Also allows saving PDFs

Safari does not support a right-click menu for extension buttons or
submenus in context menus. However, you can add a click-and-hold popover
for buttons, which is what was done here.

The menu looks like a standard OS context-menu, but is actually html
based.  Highlight color is hard-coded, so for users that have changed
highlight color in the OS that will look a bit off, but there is no CSS
magic color constant I could find.