Show browser native notifications for proxy actions
Ideally we'd like to add buttons for these notifications, so the user
can be taken to preferences to edit proxy settings and/or request
whether they want to associate new proxies. However, the
Notification.actions API is only available for Chrome at the moment.

The verbosity of the notifications has been reduced in relation to those
on firefox, because the text gets truncated with long messages.