Don't trigger button 1 on Esc key from modal prompt before click on page
Since the default button is focused initially, pressing the event from
the Esc keypress had button 1 as its target. If you clicked elsewhere in
the page first and cleared the focus, there would be no
event.target.name and the Esc key would register as button 0.

Instead, just don't pass an event to onClose for Esc and have that
always register as button 0.