Set minimum Chrome version to 55
I don't know what version the connector actually needs -- something
after 49 --  but XP and 10.6-10.8 users would be stuck on 49 and
everyone else should be (and appears to mostly be) on 57, so 55 from
December seems like a decent compromise. The main point of this is to
make it clearer to people on 49 why it's not working.