Move context-menu save options into a submenu in Chrome and Firefox
Chrome and Firefox are limited to 6 top-level menu items (including
separators). We use three on any page with a translator (primary + two
webpage options), and in Firefox we have to burn two more for the
Preferences entry and its separator, so on sites with multiple
translators we can run out of slots and end up missing some, and we
don't have room for other features.

This moves the save options into a submenu except when there's just a
PDF with no translators.

I played with conditionally showing the options directly as long as we
were under the limit, but that makes for a less-consistent experience,
and we're going to want the additional slots for other
connector functionality.

I also tried showing the primary save option + an "Other" menu so that
you could always quickly see the active translator without waiting for a
tooltip or mousing over the menu, and that worked OK, but it still takes
an extra slot and is a bit harder to parse when you're trying to choose
between multiple available translators, so I'm not sure it's worth it.

We probably want to do this in Safari too, though I'm not sure if
there's an item limit there.