Fix mouseenter/mouseleave when hovering over target selector popup
Not sure when this stopped working, but with mouseenter/mouseleave on
the iframe they now only trigger when over a 1px line at the bottom edge
of the popup. Might be a better way to fix that, but for now just have
the events on an element in the inner document and send messages to the
main page.