Use and require ImageMagick 6 for making Safari icons
ImageMagick 7 changes how channels work, so until we figure out an
equivalent command, just use ImageMagick 6.

This hardcodes a homebrew install path for the ImageMagick executable,
because it no longer really makes sense to build the Safari extension on
a non-Mac system when the Safari App Extension needs to be built on
macOS. ImageMagick can be installed with `brew install imagemagick@6`.