Use background page for all HTTP requests
Browsers have been progressively locking down HTTP requests from content
scripts, so we've moved requests to the background page when we know
they're cross-origin. Unfortunately, we can't detect ahead of time if a
request that looks like it will be same-origin redirects to another
domain, causing it to fail due to CORS/CORB [1]. While we could allow
translators to specify when a same-domain request might be redirected
across domains, it's easier to just do all HTTP requests via the
background page. I assume this is slightly less efficient, but it
probably doesn't matter in practice (and we exchange all sorts of
messages with the background page already).

(This addresses an issue reported on the dev list in a now-deleted
message about links such as http://theses.fr/2016ECDN0016/document that
can redirect to a different domain.)

[1] https://www.chromium.org/Home/chromium-security/extension-content-script-fetches