Fix translator resetting on 0 translators with Zotero closed
If Zotero wasn't open when `Zotero.Translators.init()` was called and no
translators were available, a remote repo request would be made, but
only with the last-check timestamp, so very few or no translators would
be returned and the problem wouldn't be fixed. This change forces the
remote repo request to download all metadata and reinitialize all
translators.

This will hopefully fix the vast majority of cases of people ending up
with only webpage translation, but we still need to figure out why it's
happening to begin with. When I encountered this, `translatorsMetadata`
was set properly in storage, but for some reason it wasn't returned by
`Zotero.Prefs.get()` in `Zotero.Translators.init()` (called from
`getWebTranslatorsForLocation()`).