Wait for Zotero.initDeferred.promise in getWebTranslatorsForLocation()
We are seeing reports of occasional intermittent issues with translators
not loading correctly and the reset all the way from 1af9f5d94 not fixing
it either. Dan has been able to get his connector in this state and
noticed that translator metadata was actually available via prefs,
but translators had failed to load, leading to an assumption of race
condition between Prefs.init and getWebTranslatorsForLocation() calls
(which originate from injected pages).