Remove warning for Zotero for Firefox upgraders in first-click dialog
This removes the language about "important changes to how Zotero works
in Firefox" (#346), but it leaves the dialog as Firefox-only for now. If
we make it appear in all browsers, it will appear for all existing
non-Firefox users. It wouldn't be the end of the world to encourage
people to download the app if they haven't, but it'd be better not to
bother existing users.

Before making this appear for everyone, it would also be nice to improve
the appearance of the dialog (e.g., by adding a large Zotero logo). See
also: #67, #131