Do not load hiDPI images for translators
After rebuilding typeSchemaData in
https://github.com/zotero/utilities/commit/d1e59c062b5c387f1394919dada460af162d84a2
Safari started loading hiDPI images for translators, but
it significantly breaks the icon by making the image appear
in the lower-left 1/4 of the button. I could not find a way
to fix it. There is one mention of this issue on SO but the
person ended up rebuilding their whole app[1] so we are
disabling hiDPI icons completely for now.

[1] https://stackoverflow.com/questions/71503393/safari-extension-icon-become-very-small-in-macos-12-3