Manifest V3 megacommit
Adds a build version for Chrome's Manifest V3 version of browser
extensions. Some key points:
- Manifest V3 background service workers are supposed to not run for
longer than 5 minutes to save memory and cpu cycles (or to increase CPU
usage by continuously restarting the worker). We are using a known
workaround to prevent it from being killed:
https://stackoverflow.com/questions/66618136/persistent-service-worker-in-chrome-extension
See discussion here for why it's unlikely to ever be "fixed":
https://groups.google.com/a/chromium.org/g/chromium-extensions/c/nAvia8r1cvo
- For the event that the service worker gets killed anyway, we are using
the browser.storage.session storage to store logged errors and tab state
(detected translators) so that on service worker restart things work as
expected.
- Translation had to be moved out of the page and into a sandboxed frame
which is the only place where evals are allowed in MV3. This required
modularizing the translation framework first, then setting it up in a
way that content-script code could interact with the sandbox-frame
translate in a transparent way. The communication between content script
and the sandboxed frame is done via message passing. Need testing
translators that interact with the page for a final translation result.
- The mv2 version of single-file was replaced with a mv3 compatible
version and snapshot saving seems to work fine for our needs.
- Import file interception is currently not great. DeclarativeNetRequest
does not allow filtering via headers (for content-type), but even if it
did, there is no way to attach handlers to DNR rules. The least bad way
to handle it would be by redirecting to some internal extension page with a
prompt for import, and then redirecting back to the resource if the user
decides to cancel, but that's rather intrusive when compared to staying
on the page and having the download initiated by the browser download
manager. The current implementation just allows the browser to handle
the download and also displays a prompt to import with Zotero, but we
might just as well disable that.
- Proxies not tested, but they required minimal changes and should
work fine.
- All shared code is MV2 compatible so that we don't have to maintain
multiple separate versions and tested to work, but as with all big
changes, will need to be deployed as Beta first before we're sure
nothing got broken.

Potential issues:
- There's a widely reported bug that sometimes makes service workers get
bust and not respond to any events after restart. I've encountered it
myself a few times.  Restarting the browser or disabling and reenabling
the extension fixes it. Presumably eventually addressed on Chrome's side.
- I imagine if the service worker gets killed in a middle of a HTTP
request to Zotero during an integration operation, when the citation
dialog is being displayed (during this time the extension is waiting on
a HTTP response from Zotero), integration would get stuck and Zotero
would need a restart. No idea how likely or common this might be.