Update translator refresh code
- Update translator even if stored lastUpdated is ahead of repo to
  prevent hash mismatch
- Sort translators by ID before calculating the hash

I am not sure what's that about, but I have a bunch of Connector cached
translators with weird lastUpdated dates that are ahead of what is in
our translator repo and have "00" for seconds. There is no place in our
code where lastUpdated is changed individually, and I suspect there may
be something funny going on with the way localstorage serialized that,
but still it's super weird, since you would have to parse the date,
change it, then reparse.

Also, it appears that the order of _translators is unstable since
the latest updates to cache translator metadata
and the translator hash to be checked against Zotero is calculated
without sorting the translators first, so we are getting constant
mismatches and refetching of all translator metadata on every page
navigation. Bad!