vpat 53: fix iframe not scrollable when zoomed in
One might need to zoom onto the frame to see details
better and if they zoom in too much, the bottom half
will be outside of the viewport and with no way to
scroll to it. This adds a few css tweaks to make the
frame itself scrollable if one zooms in so that it does
not fully fit into the viewport.

Relevant WCAG criterion: https://www.w3.org/WAI/WCAG21/Understanding/reflow.html