MV3: Move translation from sandboxed content iframe to offscreen (#527)
Offscreen page doesn't support eval, so we are actually embedding a
sandbox iframe there which then supports eval so we can run the
translate architecture there.

Message passing is technically complicated:
	Content Scripts <-chrome.runtime.sendMessage->
	Background Service Worker (background-worker.js) <-MessagePort->
	Offscreen Sandbox Iframe (offscreenSandbox.html)

The MessagePort is created by Offscreen Page (offscreen.html) and the
two ports are passed respectively to to Offscreen Sandbox Iframe and Background
page after which offscreen.html only performs the function of embedding
the Offscreen Sandbox and reinitializing the MessagePort in case the
Background Service Worker gets restarted by the browser.

We are moving translate to offscreen sandbox because the sandbox iframe
on content pages are causing issues on some websites where their pages
refuse to load due to loading code finding unexpected elements.

Also:
Added a wrapper to Zotero.Translate.Web which makes translate only be
concerned with scrapping the webpage (i.e. the stuff we need to be able
to run eval for). Scrapped item saving to Zotero and UI notifications
are now handled independently.

This was done because
1. The Zotero Translate architecture has no business in doing that stuff
   anyway
2. If ItemSaver runs in the Offscreen Translate page it then needs to
   be able to independently send progress callback messages to the tab
   where translation occurs, but we have no way to pass the tab
   information with which to communicate to ItemSaver if it is managed
   by Zotero.Translate, without changing the translate repo. Passing
   around that information via Zotero.Translate is dumb and creates even
   more insane and difficult to track interdependencies in translate
   architecture.

As part of this some refactoring to Zotero.Inject also occurred as that
namespace has troubled me for a while as more and more random unrelated
hard-to-place functions got crammed there. All page saving related
functionality has been moved to pageSaving.js