implement tags autocompletion
- add a new <TagsInput> component that handles the
tags input as well as renders selected tags above it
- if the Zotero desktop app passes a list of tags,
they will be used to popupulate an autocomplete popup
appearing below the tags input that will refresh as
the user types
- arrow up/down from the input will navigate the
selected option from the popup. Enter will add the selected tag
- Enter keypress on an input that does not match any
autocomplete suggestions will add the input value as
a tag
- selected tags can be removed by clicking on the X
button or navigated with right/left arrow and removed
on space keypress
- if the desktop app does not pass any tags (e.g. older
version of Zotero), on blur, tags input will be split
by comma and added as tags for backwards compatibility
- if the target collection is switched to one from
another library, selected tags are cleared, since the
new library may have completely different tags
- when autocomplete popup appears, some extra padding
is added at the bottom of <TagsInput> component
which forces the iframe to resize ensuring there is
enought space for the popup.

Fixes: zotero#310