Always resave if different translator/mode
Fixes error if saving from the button and then the context menu or vice
versa results in a different translator/mode (e.g., choosing "Web Page
without Snapshot" from the menu and then clicking the button to trigger
"Embedded Metadata" or "Web Page with Snapshot")

Years ago in dec4808602c I added an explicit `resave` flag when choosing
a translator from the context menu, but I apparently forgot to add it to
the webpage options for non-Safari browsers, which meant that saving via
the button and then choosing "Web Page without Snapshot" would fail.
But clicking the button second would fail anyway, since that wouldn't
trigger a resave. This removes the explicit flag from the menu options
and just considers it a resave if the translator/mode is different,
which also means the popup just reopens if you select the same option
from the menu after clicking the button.