send tags as an array instead of a string
That handles the edge case of tags that may contain
an actual "," character.

Decide which format to send tags in based on the supportsTagsAutocomplete
pref sent by Zotero. Minor refactoring to handle if tags are
being split by , or not when they are being added.

Also, always add tags on blur if something is typed
for consistency with other autocompelete fields.