add button to cancel/undo session saving
When the session is cancelled on button click:
- ItemSaver is marked as cancelled. Further attempts
to save/download items in SaveItems will throw an error.
- Background Zotero.ItemSaver will receive the sessionID
of the cancelled session, and will also throw an error
before an attachment is about to be downloaded and
before downloaded data is about to be sent to Zotero.
- Zotero.HTTP will cancel all pending cancellable requests.
These requests are marked with `cancellable: true`
and are used to download attachments.
- /sessionCancel endpoint will be hit, so that all subsequent
attempts to call endpoints to add items or attachments
get a 409 response.
- stylesheet refactoring to position the done and cancel
buttons in the top right