try to set missing attachment mimeType
If translators do not provide the mimeType of an attachment,
try to set it based on the content type header or, if
that is not available, based on the url.
This makes it more likely that such an attachment will be saved
with the correct attachmentContentType, which is used to determine
how the file is previewed in the itemPane.

Followup to https://github.com/zotero/zotero-connectors/pull/568